Optimize the structure
