+++ .gitignore
@@ -3,3 +3,4 @@ log*
 *.ipynb
 *.pyc
 *.swp
+/data/
